refactor: components to use journey create hook (#1459)
