+++ apps/journeys-admin/src/components/JourneyList/AddJourneyFab/index.ts
@@ -1,0 +1,1 @@
+export { AddJourneyFab } from './AddJourneyFab'
